% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1.R
\name{tab1}
\alias{tab1}
\alias{tab1.data.frame}
\alias{tab1.list}
\alias{tab1.default}
\title{tab1}
\usage{
tab1(x, sort = 1, digits = 2, ...)

\method{tab1}{data.frame}(x, sort = 1, digits = 2, ..., as_df = TRUE)

\method{tab1}{list}(x, sort = 1, digits = 2, ..., as_df = TRUE)

\method{tab1}{default}(x, sort = 1, digits = 2, ...)
}
\arguments{
\item{x}{the thing to be tabulated}

\item{sort}{should}

\item{digits}{the number of digits to round percentages to}

\item{...}{further optional arguments passed through to methods}
}
\value{
Returns a data frame with columns
 \code{value} (the values tabulated by),
 \code{count} (the number of times a value occured),
 \code{pct} (the percentage that value occured excluding NAs) and
 \code{pct_all} (the percantage that value occure including NAs)
}
\description{
tab1

tab1.data.frame

tab1.list

tab1.default
}
