% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabit.R
\name{tabit}
\alias{tabit}
\alias{tabit.data.frame}
\alias{tabit.list}
\alias{tabit.default}
\alias{ti_tab1}
\title{tabit}
\usage{
tabit(x, sort = 1, digits = 2, ...)

\method{tabit}{data.frame}(x, sort = 1, digits = 2, ...,
  as_df = TRUE)

\method{tabit}{list}(x, sort = 1, digits = 2, ..., as_df = TRUE)

\method{tabit}{default}(x, sort = 1, digits = 2, ...)

ti_tab1(x, sort = 1, digits = 2, ...)
}
\arguments{
\item{x}{the thing to be tabulated}

\item{sort}{should}

\item{digits}{the number of digits to round percentages to}

\item{...}{further optional arguments passed through to methods}

\item{as_df}{if more than one vector has been tabulated at once, should
the results be combined into a data.frame before returning}
}
\value{
Returns a data frame with columns
 \code{value} (the values tabulated by),
 \code{count} (the number of times a value occured),
 \code{pct} (the percentage that value occurred excluding NAs) and
 \code{pct_all} (the percentage that value occurred including NAs)
}
\description{
tabit

tabit.data.frame

tabit.list

tabit.default
}
\examples{

tabit(mtcars$cyl)
tabit(mtcars[, c("cyl", "am")])


}
