% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabit_x.R
\name{tabit_x}
\alias{tabit_x}
\alias{tabit_x.data.frame}
\alias{tabit_x.grouped_df}
\alias{tabit_x.default}
\title{tabit_x}
\usage{
tabit_x(x, ..., sort = 1, digits = 2, useNA = TRUE)

\method{tabit_x}{data.frame}(x, ..., sort = 1, digits = 2, useNA = TRUE)

\method{tabit_x}{grouped_df}(x, ..., sort = 1, digits = 2, useNA = TRUE)

\method{tabit_x}{default}(x, ..., sort = 1, digits = 2, useNA = "always")
}
\arguments{
\item{x}{the thing to be tabulated}

\item{...}{further optional arguments passed through to methods}

\item{sort}{should}

\item{digits}{the number of digits to round percentages to}

\item{useNA}{whether ro us NAs as value category or not, defaults to TRUE}
}
\value{
Returns a data frame with columns
 \code{variable} (the values tabulated by),
 \code{count} (the number of times a value occurred),
 \code{pct} (the percentage that value occurred excluding NAs) and
 \code{pct_all} (the percentage that value occurred including NAs)
}
\description{
tabit_x

tabit_x.data.frame

tabit_x.grouped_df

tabit_x.default
}
\examples{

tabit_x(mtcars)
tabit_x(mtcars[, c("cyl", "am")])


}
