% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1.R
\name{ti_tab1}
\alias{ti_tab1}
\alias{ti_tab1.data.frame}
\alias{ti_tab1.list}
\alias{ti_tab1.default}
\title{ti_tab1}
\usage{
ti_tab1(x, sort = 1, digits = 2, ...)

\method{ti_tab1}{data.frame}(x, sort = 1, digits = 2, ...,
  as_df = TRUE)

\method{ti_tab1}{list}(x, sort = 1, digits = 2, ..., as_df = TRUE)

\method{ti_tab1}{default}(x, sort = 1, digits = 2, ...)
}
\arguments{
\item{x}{the thing to be tabulated}

\item{sort}{should}

\item{digits}{the number of digits to round percentages to}

\item{...}{further optional arguments passed through to methods}

\item{as_df}{if more than one vector has been tabualted at once, should
the results be combined into a data.frame before returning}
}
\value{
Returns a data frame with columns
 \code{value} (the values tabulated by),
 \code{count} (the number of times a value occured),
 \code{pct} (the percentage that value occured excluding NAs) and
 \code{pct_all} (the percantage that value occure including NAs)
}
\description{
ti_tab1

ti_tab1.data.frame

ti_tab1.list

ti_tab1.default
}
\examples{

ti_tab1(mtcars$cyl)
ti_tab1(mtcars[, c("cyl", "am")])


}
