% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabit.R
\name{ti_tab1}
\alias{ti_tab1}
\alias{tabit}
\title{tabit}
\usage{
ti_tab1(...)

tabit(...)
}
\arguments{
\item{...}{parameters forwarded to tabit_1/tabit_df}
}
\description{
Tabulation function that automatically performs one way tabulations as well
as cross tabulations depending on the input type - inheriting from
data.frame or not.
}
