% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabit_1.R
\name{tabit_1}
\alias{tabit_1}
\alias{tabit_1.data.frame}
\alias{tabit_1.list}
\alias{tabit_1.default}
\title{tabit_1}
\usage{
tabit_1(x, sort = 1, digits = 2, ...)

\method{tabit_1}{data.frame}(x, sort = 1, digits = 2, ..., as_df = TRUE)

\method{tabit_1}{list}(x, sort = 1, digits = 2, ..., as_df = TRUE)

\method{tabit_1}{default}(
  x,
  sort = 1,
  digits = 2,
  useNA = "always",
  pct = TRUE,
  na_to_char = TRUE,
  ...
)
}
\arguments{
\item{x}{the thing to be tabulated}

\item{sort}{should}

\item{digits}{the number of digits to round percentages to}

\item{...}{further optional arguments passed through to methods}

\item{as_df}{if more than one vector has been tabulated at once, should
the results be combined into a data.frame before returning}

\item{useNA}{how to handle NAs, defaults to "always" to always include NA as category}

\item{pct}{add percentage columns}

\item{na_to_char}{transform NA values to "NA" chracter string or not, defaults to TRUE}
}
\value{
Returns a data frame with columns
 \code{value} (the values tabulated by),
 \code{count} (the number of times a value occured),
 \code{pct} (the percentage that value occurred excluding NAs) and
 \code{pct_all} (the percentage that value occurred including NAs)
}
\description{
tabit_1

tabit_1.data.frame

tabit_1.list

tabit_1.default
}
\examples{

tabit_1(mtcars$cyl)
tabit_1(mtcars[, c("cyl", "am")])


}
